add RNN model and merge with TCN
