+++ nn_generators/rnn/config.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+
+from nn_generators.base_config import BaseConfig
+
+
+@dataclass
+class RNNConfig(BaseConfig):
+    pass
